% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{fileinfo_local}
\alias{fileinfo_local}
\title{Get file information from local filesystem}
\usage{
fileinfo_local(path)
}
\arguments{
\item{path}{`character` path to directory}
}
\value{
`data.frame` returned by [file.info()]
}
\description{
Get file information from local filesystem
}
\examples{
\dontrun{
  fileinfo_local(".")
}
}
